Create a proxy file for all the CLI apps that use async/await
